backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
transformers>=4.40.0
sentence-transformers==2.7.0
faiss-cpu>=1.8.0
torch>=2.2.0
pydantic>=2.0.0
numpy>=1.24.0
motor>=3.3.0
python-dotenv
